refactor: move candid description
